+++ CHANGELOG.md
@@ -17,6 +17,8 @@ wind speed in knots.
 - Add flag (`plot_convention`) for windrose plots to change orientation of the
 bars.  Engineers generally want bars oriented toward the direction the wind
 is blowing toward, not from (#680).
+- Add `isolated` parameter to `geoplot.plot_values` to allow label collision
+to only be done against the plotted data in that iteration and not all data.
 - Add WPC XTEUS (national max/min temp) parser `pyiem.nws.products.xteus`.
 - Fill out `pyiem.data.reference.prodDefinitions` based on what the IEM AFOS
 database has.